put all the print stuff in one class, reduced redundant printing functions a bit.
