+++ src/tigerZone/TileCreator.java
@@ -1,0 +1,9 @@
+package tigerZone;
+
+public class TileCreator {
+	public Territory[] myTerritories = new Territory[246];
+	public TileCreator()
+	{
+		
+	}
+}
